Attempt to fix one source of blocking when Zabbix is unreachable
The "// FIXME: Move to seperate goroutine so it's non-blocking" on the select
case to fetch active checks from the Zabbix host bit us - if it doesn't return
quickly, it blocks the select for too long.  Packs stack up in the plugin's
input channel and block the Heka router.

Moved the zabbix_client.FetchActiveChecks to a goroutine (which also involved
gratuitous locking of the map).